Use vsstubs by default.

+++ make_portable.ps1
@@ -11,6 +11,7 @@ function DownloadFile {
         $FileHash = Get-FileHash -Path $OutFile -Algorithm SHA512
         if ( -Not ($FileHash.Hash -eq $Hash)) {
             Write-Output "$OutFile is broken. Delete it and try again"
+            Pop-Location
             exit
         }
     }
@@ -36,12 +37,13 @@ DownloadFile -Uri $Packages.python.url -OutFile $Packages.python.name -Hash $Pac
 DownloadFile -Uri $Packages.vapoursynth.url -OutFile $Packages.vapoursynth.name -Hash $Packages.vapoursynth.hash
 DownloadFile -Uri $Packages.vseditor.url -OutFile $Packages.vseditor.name -Hash $Packages.vseditor.hash
 # DownloadFile -Uri $Packages.vsrepogui.url -OutFile $Packages.vsrepogui.name -Hash $Packages.vsrepogui.hash
-DownloadFile -Uri $Packages.vspreview.url -OutFile $Packages.vspreview.name
+DownloadFile -Uri $Packages.vspreview.url -OutFile $Packages.vspreview.name -Hash $Packages.vspreview.hash
 DownloadFile -Uri $Packages.lexpr.url -OutFile $Packages.lexpr.name -Hash $Packages.lexpr.hash
 DownloadFile -Uri $Packages.getpip.url -OutFile $Packages.getpip.name
 DownloadFile -Uri $Packages.ocr.url -OutFile $Packages.ocr.name -Hash $Packages.ocr.hash
 DownloadFile -Uri $Packages.imwri.url -OutFile $Packages.imwri.name -Hash $Packages.imwri.hash
 DownloadFile -Uri $Packages.subtext.url -OutFile $Packages.subtext.name -Hash $Packages.subtext.hash
+DownloadFile -Uri $Packages.vsstubs.url -OutFile $Packages.vsstubs.name
 Pop-Location
 
 
@@ -67,6 +69,7 @@ Push-Location -Path downloads
 Expand-Archive -Path $Packages.'7za'.name -DestinationPath "7za" -Force
 Expand-Archive -Path $Packages.vspreview.name -DestinationPath vspreview -Force
 # Expand-Archive -Path $Packages.vsrepogui.name -DestinationPath VSRepoGUI -Force
+Expand-Archive -Path $Packages.vsstubs.name -DestinationPath vsstubs -Force
 Expand-Archive -Path $Packages.vapoursynth.name -Destination ..\VapourSynth -Force
 Expand7Zip -Path $Packages.vseditor.name -Destination ..\VapourSynth\
 Expand7Zip -Path $Packages.lexpr.name -Destination ..\VapourSynth\vapoursynth64\plugins\
@@ -80,11 +83,15 @@ Pop-Location
 $Requirements = Get-Item .\downloads\vspreview\vapoursynth-preview-$($Packages.vspreview.branch)\requirements.txt
 Set-Content -Path $Requirements (Get-Content -Path $Requirements | Select-String -Pattern 'vapoursynth' -NotMatch )
 .\VapourSynth\python.exe -m pip install -r $Requirements --no-warn-script-location
+.\VapourSynth\python.exe -m pip install .\downloads\vsstubs\VapourSynth-Plugins-Stub-Generator-$($Packages.vsstubs.branch)\vsstubs\ --no-warn-script-location
+.\VapourSynth\python.exe -m vsstubs install
 
 
-Remove-Item -Path .\VapourSynth\__pycache__ -Recurse -Force
+if ( Test-Path -Path .\VapourSynth\__pycache__ ) {
+    Remove-Item -Path .\VapourSynth\__pycache__ -Recurse -Force
+}
 Move-Item -Path .\VapourSynth\sitecustomize.py -Destination .\VapourSynth\Lib\ -Force
-Move-Item -Path .\VapourSynth\vapoursynth.cp*.pyd -Destination .\VapourSynth\Lib\ -Force
+Move-Item -Path .\VapourSynth\vapoursynth.cp*.pyd -Destination .\VapourSynth\Lib\site-packages\ -Force
 Copy-Item -Path .\downloads\vspreview\vapoursynth-preview-$($Packages.vspreview.branch)\vspreview -Destination .\VapourSynth\Lib\site-packages\ -Recurse -Force
 # Copy-Item -Path .\downloads\VSRepoGUI\VSRepoGUI.exe -Destination .\VapourSynth\ -Force
 Copy-Item -Path .\vsrepogui.json -Destination .\VapourSynth\ -Force
@@ -97,6 +104,7 @@ Push-Location -Path downloads
 Remove-Item -Path 7za -Recurse -Force
 Remove-Item -Path vspreview -Recurse -Force
 # Remove-Item -Path VSRepoGUI -Recurse -Force
+Remove-Item -Path vsstubs -Recurse -Force
 Pop-Location
 
 # Remove some extra files we don't need.
